changed legacy code
