+++ package.json
@@ -36,7 +36,8 @@
       "jsx-a11y/click-events-have-key-events": 0,
       "jsx-a11y/no-static-element-interactions": 0,
       "react/prop-types": 0,
-      "react/no-array-index-key": 0
+      "react/no-array-index-key": 0,
+      "react-hooks/exhaustive-deps": 0
     },
     "env": {
       "browser": true
