+++ src/components/Footer.js
@@ -5,7 +5,7 @@ export default function Footer() {
   return (
     <FooterStyle>
       <footer>
-        Copyright (c) Prography Prography 5기 jsa 팀 도디 승아 동원 용민
+        {/* Copyright (c) Prography Prography 5기 jsa 팀 도디 승아 동원 용민 */}
       </footer>
     </FooterStyle>
   );
