+++ src/components/index.js
@@ -3,7 +3,7 @@ export { default as Nav } from './Nav';
 export { default as ResultPage } from './ResultPage';
 export { default as ResultsPageDetail } from './ResultsPageDetail';
 export { default as ResultsPageList } from './ResultsPageList';
-export { default as SelectPageDody } from './SelectPageDody';
+export { default as SelectPage } from './SelectPage';
 export { default as MyPage } from './MyPage';
 export { default as EventList } from './EventList';
 export { default as Footer } from './Footer';
